pyproject.toml
[project]
name = "milk-link"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiomysql>=0.2",
    "alembic>=1.13",
    "apscheduler>=3.10",
    "fastapi>=0.115",
    "httpx>=0.27",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "sqlalchemy[asyncio]>=2.0",
    "uvicorn[standard]>=0.30",
]
